Update README.md
Basic program to display events from an Arduino IR sensor.  The goal is to eventually use this to track rotations of a spinning stationary bike wheel and display instantaneous speed, average speed, total distance, and elapsed time.  I'm currently using an Arduino Uno to run the IR sensor and the output is displayed on a laptop.  I would eventually like to send the output to a Raspberry Pi linked to a small LCD screen.